pyproject.toml
[tool.poetry]
name = "wyvern-ai"
version = "0.0.1"
description = ""
authors = ["Wyvern AI <info@wyvern.ai>"]
readme = "README.md"
packages = [
    { include = "wyvern" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pydantic = "^1.10.4"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
typer = {extras = ["all"], version = "^0.9.0"}
pyyaml = "^6.0"
pyhumps = "^3.8.0"
python-dotenv = "^1.0.0"
pandas = "1.5.3"
feast = {extras = ["redis", "snowflake"], version = "^0.31.1"}
httpx = "^0.24.1"
snowflake-connector-python = "3.0.3"
boto3 = "^1.26.146"
ddtrace = "^1.14.0"
msgspec = "^0.16.0"
lz4 = "^4.3.2"
more-itertools = "^9.1.0"
tqdm = "^4.65.0"
nest-asyncio = "^1.5.7"


[tool.poetry.group.dev.dependencies]
ipython = "^8.9.0"
pytest = "^7.2.1"
isort = "^5.12.0"
types-pyyaml = "^6.0.12.6"
black = "^22.6.0"
pip-tools = "^6.12.2"
twine = "^4.0.2"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
types-boto3 = "^1.0.2"
pyinstrument = "^4.4.0"
pytest-dotenv = "^0.5.2"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module=[
    "sentence_transformers.*",
    "scipy.spatial.distance.*",
    "featureform.*",
    "setuptools.*",
    "ddtrace.*",
    "nest_asyncio.*",
    "lz4.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-v"
filterwarnings = ["ignore::DeprecationWarning"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
env_files = [
    ".testing_env",
]
